Added file nighttime-timer
Using CC computer, this sends RS signal to device to only operate from 8:00PM to just after 6:00AM

It is not setup with variables. Edit "back" to the side of the CC computer you want to send signal from.
The time used is MC time from os.time(). 
8:00PM is 20:00 in 24hr time, which is equal to 20.000
6:06AM is 06:06 which is 6.100. I chose the "6.100" first, so it would be after 6AM. 0.100 of one hour works out to 6 minutes, so the timer goes off at 6:06AM
Some more examples: 18.200 is 6:12PM, 12.500 is 12:30PM, 0.700 is 12:42AM
Adjust these times as needed.